refactor(step3):  view name
view 클래스 이름에 쓸때없는 단어 삭제하여 쉽게 알아볼수있게 수정 (Lotto라는 단어가 필요가없음)
view에서도 역할을 구분하기위해 input, result  둘로 나눔
Scanner 재사용
